+++ Example6.SingleLevelSharding/Services/AddTenantNameClaim.cs
@@ -21,7 +21,7 @@ public class AddTenantNameClaim : IClaimsAdder
         _userAdmin = userAdmin;
     }
 
-    public async Task<Claim> AddClaimToUserAsync(string userId)
+    public async Task<Claim?> AddClaimToUserAsync(string userId)
     {
         var user = (await _userAdmin.FindAuthUserByUserIdAsync(userId)).Result;